Add test cases and delegated logic
