+++ controllers/datadogagent/controller_test.go
@@ -2832,7 +2832,7 @@ func createClusterAgentDependencies(c client.Client, dda *datadoghqv1alpha1.Data
 // the metricForwardersManager logic is tested in the util/datadog package
 type dummyManager struct{}
 
-func (dummyManager) Register(datadog.MonitoredObject) {
+func (dummyManager) Register(obj datadog.MonitoredObject, objectKind string) {
 }
 
 func (dummyManager) Unregister(datadog.MonitoredObject) {
